+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Use single API call to identify payment sender.
 - Migrate build system to Vite 3.
+- Use different color when rendering external links.
 
 ### Fixed